+++ client/scripts/views/components/sidebar/index.ts
@@ -15,7 +15,6 @@ import { ProposalType } from 'identifiers';
 import { link } from 'helpers';
 import { ChainClass, ChainBase, ChainNetwork, ChainInfo, CommunityInfo, AddressInfo, NodeInfo } from 'models';
 import NewTopicModal from 'views/modals/new_topic_modal';
-import EditTopicModal from 'views/modals/edit_topic_modal';
 
 import ChainStatusIndicator from 'views/components/chain_status_indicator';
 import { MobileNewProposalButton } from 'views/components/new_proposal_button';
@@ -98,61 +97,6 @@ const OffchainNavigationModule: m.Component<{ sidebarTopic: number }, { dragulaI
     const onSearchPage = (p) => p.startsWith(`/${app.activeId()}/search`);
     const onChatPage = (p) => p === `/${app.activeId()}/chat`;
 
-    const featuredTopics = {};
-    const otherTopics = {};
-    const featuredTopicIds = app.community?.meta?.featuredTopics || app.chain?.meta?.chain?.featuredTopics;
-
-    const getTopicRow = (id, name, description) => m(ListItem, {
-      class: 'topic-row',
-      key: id,
-      contentLeft: m('.proposal-topic-icon'),
-      contentRight: m.route.get() === `/${app.activeId()}/discussions/${encodeURI(name)}`
-        && app.user.isAdminOfEntity({ chain: app.activeChainId(), community: app.activeCommunityId() })
-        && m(PopoverMenu, {
-          class: 'sidebar-edit-topic',
-          position: 'bottom',
-          transitionDuration: 0,
-          hoverCloseDelay: 0,
-          closeOnContentClick: true,
-          trigger: m(Icon, {
-            name: Icons.CHEVRON_DOWN,
-          }),
-          content: m(MenuItem, {
-            label: 'Edit topic',
-            onclick: (e) => {
-              app.modals.create({
-                modal: EditTopicModal,
-                data: { description, id, name }
-              });
-            }
-          })
-        }),
-      label: [
-        name,
-      ],
-      active: m.route.get() === `/${app.activeId()}/discussions/${encodeURI(name.toString().trim())}`
-        || (sidebarTopic && sidebarTopic === id),
-      onclick: (e) => {
-        e.preventDefault();
-        m.route.set(`/${app.activeId()}/discussions/${name}`);
-      },
-    });
-
-    app.topics.getByCommunity(app.activeId()).forEach((topic) => {
-      const { id, name, description } = topic;
-      if (featuredTopicIds.includes(`${topic.id}`)) {
-        featuredTopics[topic.name] = { id, name, description, featured_order: featuredTopicIds.indexOf(`${id}`) };
-      } else {
-        otherTopics[topic.name] = { id, name, description };
-      }
-    });
-    const otherTopicListItems = Object.keys(otherTopics)
-      .sort((a, b) => otherTopics[a].name.localeCompare(otherTopics[b].name))
-      .map((name, idx) => getTopicRow(otherTopics[name].id, name, otherTopics[name].description));
-    const featuredTopicListItems = Object.keys(featuredTopics)
-      .sort((a, b) => Number(featuredTopics[a].featured_order) - Number(featuredTopics[b].featured_order))
-      .map((name, idx) => getTopicRow(featuredTopics[name].id, name, featuredTopics[name].description));
-
     return m('.OffchainNavigationModule.SidebarModule', [
       m(List, [
         m(ListItem, {
@@ -179,7 +123,7 @@ const OffchainNavigationModule: m.Component<{ sidebarTopic: number }, { dragulaI
           active: onDiscussionsPage(m.route.get())
             && (app.chain ? app.chain.serverLoaded : app.community ? app.community.serverLoaded : true)
             && (sidebarTopic === null || !m.route.get().startsWith(`/${app.activeId()}/proposal/discussion/`)),
-          label: 'All Discussions',
+          label: 'Discussions',
           onclick: (e) => {
             e.preventDefault();
             m.route.set(`/${app.activeId()}`);
@@ -206,28 +150,6 @@ const OffchainNavigationModule: m.Component<{ sidebarTopic: number }, { dragulaI
         //   contentLeft: m(Icon, { name: Icons.MESSAGE_CIRCLE }),
         // }),
       ]),
-      m(List, [
-        featuredTopicListItems.length === 0 && otherTopicListItems.length === 0 && !app.threads.initialized
-          && m(ListItem, {
-            class: 'section-callout',
-            label: m('div', { style: 'text-align: center' }, m(Spinner, { active: true, size: 'xs' })),
-          }),
-      ]),
-      m(List, {
-        onupdate: (vvnode) => {
-          if (app.user.isAdminOfEntity({ chain: app.activeChainId(), community: app.activeCommunityId() })
-              && !vnode.state.dragulaInitialized) {
-            vnode.state.dragulaInitialized = true;
-            dragula([vvnode.dom]).on('drop', async (el, target, source) => {
-              const reorder = Array.from(source.children).map((child) => {
-                return (child as HTMLElement).id;
-              });
-              await app.community.meta.updateFeaturedTopics(reorder);
-            });
-          }
-        }
-      }, featuredTopicListItems),
-      m(List, { class: 'more-topics-list' }, otherTopicListItems),
     ]);
   }
 };
