+++ client/styles/pages/discussions/index.scss
@@ -19,6 +19,39 @@
         }
     }
 
+    .discussions-topics {
+        margin-bottom: 20px;
+        font-size: 17px;
+        line-height: 1.25;
+    }
+    .discussions-topic {
+        color: $text-color-medium;
+        display: inline-block;
+        margin-right: 24px;
+        cursor: pointer;
+        @include proposal-topic-inline;
+        &:hover .proposal-topic-name {
+            text-decoration: underline;
+        }
+        .sidebar-edit-topic {
+            display: inline-block;
+        }
+        .proposal-topic-icon {
+            border-color: $text-color-medium;
+            margin-right: 8px;
+        }
+        .proposal-topic-name,
+        .proposal-topic-icon {
+            display: inline-block;
+        }
+        &.active {
+            color: $text-color-black;
+            .proposal-topic-icon {
+                border-color: $text-color-black;
+            }
+        }
+    }
+
     // content
     .LastSeenDivider {
         display: flex;
