+++ client/styles/sublayout.scss
@@ -30,10 +30,6 @@
                 font-size: 1.25rem;
                 margin: 7px 17px 0 0;
             }
-            .sublayout-header-description {
-                margin-top: 2px;
-                line-height: 1.1;
-            }
         }
         .sublayout-header-right {
             display: flex;
